refactor(dispatch_contract): split 699-line file into dispatch/ subpackage (closes #362)
Split backend/dispatch_contract.py into five focused submodules:

- dispatch/signing.py   — HMAC-SHA256 sign_payload/verify_payload, timestamp validation,
                          internal _sign_envelope_payload/_verify_envelope_signature helpers.
                          Designed to be reusable by Maxwell-Daemon via shared_scripts/.
- dispatch/registry.py  — DispatchAccess, DispatchAction, ALLOWLISTED_ACTIONS,
                          get_action(), requires_confirmation(), _scheduler_modify_command()
- dispatch/envelope.py  — CommandEnvelope, DispatchConfirmation, schema constants
                          (SCHEMA_VERSION, ENVELOPE_VERSION, SUPPORTED_SCHEMA_VERSIONS)
- dispatch/validate.py  — validate_envelope(), validate_envelope_crypto(),
                          DispatchValidationResult, CryptoValidationResult
- dispatch/audit.py     — DispatchAuditLogEntry, build_audit_log_entry()
- dispatch/__init__.py  — re-exports full public API for ergonomic imports
- dispatch/SCHEMA_MIGRATIONS.md — v1→v2 transition rules

backend/dispatch_contract.py is now a thin compatibility shim (~65 lines) that
re-exports everything; all existing imports continue to work without changes to
any caller.

All tests in test_dispatch_contract.py and test_envelope_signing.py pass.
Ruff check and format clean.
